requirements.txt
# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0

# Image processing
Pillow>=9.5.0
opencv-python>=4.7.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0

# Training utilities
accelerate>=0.20.0
wandb>=0.15.0
tqdm>=4.65.0

# Evaluation and metrics
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-metrics>=0.1.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
jupyter>=1.0.0
ipywidgets>=8.0.0
